Making updates related to fog api changes.
If we try to use 'endpoint' or 'compute_url' it will just silently
default to dfw and never create anything in ord.

Changed instances of endpoint to rackspace_region to be consistent
with fog.

Note that a bit of the documentation refers to the :lon region, but
this is not useable without the ability to set the auth url. This
was addressed in https://github.com/mitchellh/vagrant-rackspace/pull/4
but that had a couple other changes as well.