Switch to WaitForCommunicator
The OpenStack provider's action_up used a custom WaitForServerToBeAccessible
action to poll instances via SSH in order to determine when the VM is up and
running. One large shortcoming of this custom implementation is that it does
not support communicators other than SSH, notably WinRM. This patch replaces
the WaitForServerToBeAccessible with the built-in WaitForCommunicator action
which has been part of Vagrant since v1.3 and is able to handle multiple
communicators.

Fixes ggiamarchi/vagrant-openstack-provider#264.
Closes ggiamarchi/vagrant-openstack-provider#227.