Switch to Vagrant SyncedFolders
Vagrant 1.5 introduced a standard SyncedFolders middleware which can be re-used
by providers to implement things such as rsync support. This patch uses Vagrant
SyncedFolders as the default sync implementation instead of the legacy
SyncFolder implementation inherited from the Rackspace project.

The legacy implementation is retained for backwards compatibility and will be
used if any of the following provider options is set:

  - rsync_includes
  - rsync_ignore_files
  - sync_method

A deprecation warning will be printed.