Enable parallel actions
The `parallel: true` flag is added to the provider plugin to tell Vagrant that
actions such as `vagrant up` on multiple machines can be run in parallel. This
works almost out of the box, the only hickup is the search for free floating IPs
which needs to be made thread safe using a `Mutex` object.

I've tested this on a four machine, with different network settings and a wide
range of different provisioners, including:
- Shell
- File
- Docker
- Ansible

It all worked fine and resulted in a much faster CI pipeline.

Closes #130