Discover URLs instead of constructing in neutron
The approach of constructing URLs via string concatenation is brittle
and breaks with projects implementing the openstack API but advertising
URLs with trailing slash. In that case URLs contains two consecutives
slashes ending up in the plugin not working. Instead, discover the
required URLs by parsing the top level response and choosing the URL to
use