Fix deadlock for parallel `vagrant up`
When some floating IPs are allocated but unused and when
floating_ip_pool_always_allocate = false this can lead
into a situation where several servers are trying to use
the same IP address and it leads to a deadlock because
they all wait to have the IP assigned and of course only
one of them can have it.

The mutex has to be at a higher level than it was to
ensure floating IP assignment to be well done in an
atomic way.

Of course, doing that, floating IP assignment does not
occur in parallel anymore but it looks to be the only
way to solve the problem.