Merge configuration volume_boot hashes
This commit updates the merge behavior of the config object to merge
volume_boot hashes together. This allows defaults to be set for all
VMs that are booting from a volume:

    Vagrant.configure('2') do |config|

      # Always delete volumes when VMs are destroyed.
      config.vm.provider :openstack do |os|
        os.volume_boot = {delete_on_destroy: true}
      end

      config.vm.define 'centos1' do |node|
        node.vm.provider :openstack do |os|
          os.volume_boot = {image: 'centos-7.5',
                            size: 20}
        end
      end
    end

Prior to this change, the last `volume_boot` hash specified would overwrite
all configuration set by other provider blocks.